--- controllers/pegawai.controller.go
@@ -1,66 +1,0 @@
-package controllers
-
-import (
-	"net/http"
-	"prakerja/models"
-	"strconv"
-
-	"github.com/labstack/echo/v4"
-)
-
-func FetchAllPegawai(c echo.Context) error {
-	result, err := models.FetchAllPegawai()
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, map[string]string{"message": err.Error()})
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
-
-func StorePegawai(c echo.Context) error {
-	nama := c.FormValue("nama")
-	email := c.FormValue("email")
-	telepon := c.FormValue("telepon")
-
-	result, err := models.StorePegawai(nama, email, telepon)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, map[string]string{"message": err.Error()})
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
-
-func UpdatePegawai(c echo.Context) error {
-	id := c.FormValue("id")
-	nama := c.FormValue("nama")
-	email := c.FormValue("email")
-	telepon := c.FormValue("telepon")
-
-	conv_id, err := strconv.Atoi(id)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	result, err := models.UpdatePegawai(conv_id, nama, email, telepon)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
-
-func DeletePegawai(c echo.Context) error {
-	id := c.FormValue("id")
-
-	conv_id, err := strconv.Atoi(id)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	result, err := models.DeletePegawai(conv_id)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
