+++ models/employees/employee.go
@@ -1,0 +1,8 @@
+package employees
+
+type Pegawai struct {
+	Id      int    `json:"id"`
+	Nama    string `json:"nama"`
+	Email   string `json:"email"`
+	Telepon string `json:"telepon"`
+}
