--- models/pegawai.model.go
@@ -1,139 +1,0 @@
-package models
-
-import (
-	"net/http"
-	"prakerja/db"
-)
-
-type Pegawai struct {
-	Id      int    `json:"id"`
-	Nama    string `json:"nama"`
-	Email   string `json:"email"`
-	Telepon string `json:"telepon"`
-}
-
-func FetchAllPegawai() (Response, error) {
-	var obj Pegawai
-	var arrobj []Pegawai
-	var res Response
-
-	con := db.CreateCon()
-
-	sqlStatement := "SELECT * FROM pegawai"
-
-	rows, err := con.Query(sqlStatement)
-
-	if err != nil {
-		return res, err
-	}
-
-	defer rows.Close()
-
-	for rows.Next() {
-		err = rows.Scan(&obj.Id, &obj.Nama, &obj.Email, &obj.Telepon)
-		if err != nil {
-			return res, err
-		}
-
-		arrobj = append(arrobj, obj)
-	}
-
-	res.Status = http.StatusOK
-	res.Message = "Success"
-	res.Data = arrobj
-
-	return res, nil
-}
-
-func StorePegawai(nama string, email string, telepon string) (Response, error) {
-	var res Response
-
-	con := db.CreateCon()
-
-	sqlStatement := "INSERT pegawai (nama, email, telepon) VALUES (?, ?, ?)"
-
-	stmt, err := con.Prepare(sqlStatement)
-	if err != nil {
-		return res, err
-	}
-
-	result, err := stmt.Exec(nama, email, telepon)
-	if err != nil {
-		return res, err
-	}
-
-	lastInsertedId, err := result.LastInsertId()
-	if err != nil {
-		return res, err
-	}
-
-	res.Status = http.StatusOK
-	res.Message = "Success"
-	res.Data = map[string]int64{
-		"last_inserted_id": lastInsertedId,
-	}
-
-	return res, nil
-}
-
-func UpdatePegawai(id int, nama string, email string, telepon string) (Response, error) {
-	var res Response
-
-	con := db.CreateCon()
-
-	sqlStatement := "UPDATE pegawai SET nama = ?, email = ?, telepon = ? WHERE id = ?"
-
-	stmt, err := con.Prepare(sqlStatement)
-	if err != nil {
-		return res, err
-	}
-
-	result, err := stmt.Exec(nama, email, telepon, id)
-	if err != nil {
-		return res, err
-	}
-
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return res, err
-	}
-
-	res.Status = http.StatusOK
-	res.Message = "Success"
-	res.Data = map[string]int64{
-		"rows_affected": rowsAffected,
-	}
-
-	return res, nil
-}
-
-func DeletePegawai(id int) (Response, error) {
-	var res Response
-
-	con := db.CreateCon()
-
-	sqlStatement := "DELETE FROM pegawai WHERE id = ?"
-
-	stmt, err := con.Prepare(sqlStatement)
-	if err != nil {
-		return res, err
-	}
-
-	result, err := stmt.Exec(id)
-	if err != nil {
-		return res, err
-	}
-
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return res, err
-	}
-
-	res.Status = http.StatusOK
-	res.Message = "Success"
-	res.Data = map[string]int64{
-		"rows_affected": rowsAffected,
-	}
-
-	return res, nil
-}
